populate data command done
